dist_import/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
Pillow==10.4.0
numpy==1.26.4
ultralytics==8.3.39
opencv-python-headless==4.10.0.84
onnxruntime==1.18.1

